+++ src/main/java/game/GameRandom.java
@@ -1,4 +1,4 @@
-package racingcar;
+package game;
 
 public interface GameRandom {
     int randomNumberRange(int start, int end);
