+++ src/main/java/game/Input.java
@@ -1,4 +1,4 @@
-package racingcar;
+package game;
 
 public interface Input {
     String readLine();
