+++ src/main/java/game/Output.java
@@ -1,4 +1,4 @@
-package racingcar;
+package game;
 
 public interface Output {
     void println(String str);
