+++ src/test/java/racingcar/FixedGameRandom.java
@@ -1,8 +1,9 @@
 package racingcar;
 
+import game.GameRandom;
 import java.util.List;
 
-public class FixedGameRandom implements GameRandom{
+public class FixedGameRandom implements GameRandom {
 
     private final List<Integer> numbers;
     private int cur = 0;
